fix: 주문/결제 페이지에서 결제 완료시, 장바구니 아이템 비우도록 로직 수정
