+++ src/pages/OrderCheckoutPage/OrderCheckoutPage.js
@@ -1,4 +1,5 @@
 import { useHistory, useLocation } from 'react-router';
+import { useDispatch, useSelector } from 'react-redux';
 import {
   Container,
   CheckoutListContainer,
@@ -11,13 +12,20 @@ import { useServerAPI } from '../../hooks';
 import { numberWithCommas } from '../../shared/utils';
 import { Header, PaymentInfoBox, RowProductItem } from '../../components';
 import ScreenContainer from '../../shared/styles/ScreenContainer';
+import { updateShoppingCartItemsAsync } from '../../redux/action';
 
 const OrderCheckoutPage = () => {
   const history = useHistory();
   const location = useLocation();
+  const dispatch = useDispatch();
+
+  const { myShoppingCartId, myShoppingCartProductIds } = useSelector(state => ({
+    myShoppingCartId: state.myShoppingCartReducer.myShoppingCart.id,
+    myShoppingCartProductIds: state.myShoppingCartReducer.myShoppingCart.productIdList,
+  }));
 
   const checkedItemList = location.state?.checkedItemList;
-
+  const checkedIdList = checkedItemList.map(item => item.id);
   const expectedPrice = checkedItemList.reduce((acc, item) => {
     const { price, amount } = item;
 
@@ -27,7 +35,12 @@ const OrderCheckoutPage = () => {
   const { postData: createOrder } = useServerAPI([], SCHEMA.ORDER);
 
   const onClickPaymentButton = () => {
-    if (!window.confirm(CONFIRM_MESSAGE.CHECKOUT)) return;
+    if (!window.confirm(CONFIRM_MESSAGE.PURCHASE)) return;
+
+    const newContent = {
+      productIdList: myShoppingCartProductIds.filter(productId => !checkedIdList.includes(productId)),
+    };
+    dispatch(updateShoppingCartItemsAsync(SCHEMA.SHOPPING_CART, myShoppingCartId, newContent));
 
     const content = {
       orderedProductList: checkedItemList.map(({ id, amount }) => ({ id, amount })),
